+++ test/resources/language/sh/shellcheck-infer.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env zsh
+# shellcheck shell=bash
+var="foo bar"
+if [[ $var =~ "foo.*bar" ]]; then
+    echo "Match!"
+fi